+++ src/enzo/PhotonGrid_Methods.h
@@ -412,7 +412,7 @@ int WalkPhotonPackage(PhotonPackageEntry **PP,
 		      float LengthUnits, float TimeUnits, float LightSpeed,
 		      float MinimumPhotonFlux);
 
-int FindPhotonNewGrid(int cindex, FLOAT *r, FLOAT *u, int *g,
+int FindPhotonNewGrid(int cindex, FLOAT *r, double *u, int *g,
 		      PhotonPackageEntry* &PP,
 		      grid* &MoveToGrid, int &DeltaLevel,
 		      const float *DomainWidth, int &DeleteMe,
