+++ src/enzo/PhotonPackage.h
@@ -52,7 +52,7 @@ public:
 
   void PrintInfo(void) {
     FLOAT r[3];
-    FLOAT u[3];
+    double u[3];
     pix2vec_nest64((int64_t)(1 << level), ipix, u);
     for (int dim = 0; dim < 3; dim++) 
       r[dim] = SourcePosition[dim] + u[dim] * Radius;
